+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Executions are paging + are implemented as lazy enumerators
 - Fixed after getter in schedule
 - Blueprint works with deprecated labels 
+- Metadata object has new methods for working with unlisted attribute
+- Metadata method deprecated= now accepts and return true/false
 
 ## 0.6.21
 - Consolidated error reporting for Domain#create_users & Project#import_users